pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "ailit-agent"
version = "0.1.0"
description = "ailit-agent: ailit_base, agent_work, agent_memory, ailit_runtime, ailit_cli, workflow engine"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27",
    "jsonschema>=4.20",
    "pyyaml>=6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
ailit = "ailit_cli.cli:main"

[tool.setuptools.packages.find]
where = ["ailit"]
include = [
    "ailit_base*",
    "agent_memory*",
    "ailit_runtime*",
    "agent_work*",
    "ailit_cli*",
    "workflow_engine*",
    "project_layer*",
]

[tool.pytest.ini_options]
pythonpath = ["ailit", "tests/e2e"]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-m 'not integration and not manual_model_e2e'"
markers = [
    "integration: сетевые live-тесты (ключ и явное согласие)",
    "e2e: CLI и сгенерированное приложение в .ailit (subprocess)",
    "manual_model_e2e: e2e/live-тесты с реальной моделью, запускать только вручную",
]

[tool.ruff]
line-length = 100
target-version = "py310"
